// src/enums/enums.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  USER = 'USER',
  BUSINESS = 'BUSINESS',
}

export enum Status {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SUSPENDED = 'SUSPENDED',
  BLOCKED = 'BLOCKED',
  CANCELED = 'CANCELED',
  RELEASED = 'RELEASED',
  FAILED = 'FAILED',
  CLOSED = 'CLOSED',
  SUCCESSFUL = 'SUCCESSFUL',
  AVAILABLE = 'AVAILABLE',
  UNAVAILABLE = 'UNAVAILABLE',
  PENDING = 'PENDING',
  INPROGRESS = 'INPROGRESS',
  FOUND = 'FOUND',
}

export enum ActionType {
  OWNERSHIPCHANGE = 'OWNERSHIPCHANGE',
  ADJUSTMENT = 'ADJUSTMENT',
  UPLOAD = 'UPLOAD',
  UPDATE = 'UPDATE',
}
export enum UnitActivity {
  DEBIT = 'DEBIT',
  CREDIT = 'CREDIT',
}
export enum ResourceType {
  SMARTDEVICE = 'SMARTDEVICE',
  VEHICLE = 'VEHICLE',
  HOUSE = 'HOUSE',
  LAND = 'LAND',
}
export enum AccountType {
  ADMIN = 'ADMIN',
  BUSINESS = 'BUSINESS',
  INDIVIDUAL = 'INDIVIDUAL',
}